fix: adjustment routes comments
