+++ backend/src/modules/comments/comments.controller.ts
@@ -20,7 +20,10 @@ import { CreateCommentDto } from './dto/create-comment.dto';
 import { UpdateCommentDto } from './dto/update-comment.dto';
 import { CommentsService } from './comments.service';
 import { CommentPermissionGuard } from './guards/updateOrDeletePermission.guard';
-import { CommentsSwagger, CommentsUserSwagger } from './swagger/comments.swagger';
+import {
+  CommentsSwagger,
+  CommentsUserSwagger,
+} from './swagger/comments.swagger';
 import { BadRequestSwagger } from 'src/helpers/swagger/bad-request.swagger';
 import { NotFoundSwagger } from 'src/helpers/swagger/not-found.swagger';
 import { UnauthorizedSwagger } from 'src/helpers/swagger/unauthorized.swagger';
@@ -96,7 +99,7 @@ export class CommentsController {
   @ApiResponse({
     status: 401,
     description: 'Você não tem permissão para alterar o comentário.',
-    type: UnauthorizedSwagger
+    type: UnauthorizedSwagger,
   })
   @UseGuards(JwtauthGuard, CommentPermissionGuard)
   update(@Param('id') id: string, @Body() data: UpdateCommentDto) {
